requirements.txt
chromadb== 0.5.3
protobuf==3.20.*
streamlit==1.36.0
pypdf==4.2.0
langchain==0.2.5
langchain-community==0.2.5
langchain-huggingface==0.0.3
python-dotenv